+++ src/somalierpkg/relate.nim
@@ -131,10 +131,19 @@ proc add_ped_samples(grouped: var seq[pair], samples:seq[Sample], sample_names:s
       else:
         grouped.add((sampleB.id, sampleA.id, rel))
 
-proc readGroups(path:string): seq[pair] =
+
+proc readGroups(path:string, existing_groups: var seq[pair]): seq[pair] =
   result = newSeq[pair]()
   if path == "":
     return
+
+  var extbl = newTable[string, seq[pair]]()
+  # seen makes sure we don't add a pair that's already present
+  var seen = newTable[tuple[a:string, b:string], bool]()
+  for ex in existing_groups:
+    extbl.mgetOrPut(ex.a, @[]).add(ex)
+    extbl.mgetOrPut(ex.b, @[]).add(ex)
+    seen[(ex.a, ex.b)] = true
 
   # expand out a,b,c to a,b, a,c, b,c
   for line in path.lines:
@@ -151,6 +160,25 @@ proc readGroups(path:string): seq[pair] =
           result.add((x, y, rel))
         else:
           result.add((y, x, rel))
+
+        var added = result[result.high]
+        for up in extbl.getOrDefault(added.a, @[]):
+          var toadd:pair = (added.b, up.b, up.rel)
+          if toadd.b < toadd.a: swap(toadd.a, toadd.b)
+          # we know up.a and added.a ar already pairs so we need to pair up.b
+          # and added.a
+          if (toadd.a, toadd.b) notin seen:
+            seen[(toadd.a, toadd.b)] = true
+            existing_groups.add(toadd)
+
+        # repeat above for b
+        for up in extbl.getOrDefault(added.b, @[]):
+          var toadd:pair = (added.a, up.b, up.rel)
+          if toadd.b < toadd.a: swap(toadd.a, toadd.b)
+          if (toadd.a, toadd.b) notin seen:
+            seen[(toadd.a, toadd.b)] = true
+            existing_groups.add(toadd)
+
 
 proc n_samples(r: relation_matrices): int {.inline.} =
   return r.samples.len
@@ -204,22 +232,24 @@ iterator relatedness(r: var relation_matrices, grouped: var seq[pair]): relation
                      )
 
 
-
 template proportion_other(c:allele_count): float =
   if c.nother == 0: 0'f else: c.nother.float / (c.nother + c.nref + c.nalt).float
 
 proc ab*(c:allele_count, min_depth:int): float {.inline.} =
-  if c.proportion_other > 0.04: return -1
+  ## get the allele balance for the allele_count object while requing a min depth
+  # allow high-ish proportion other see: 
+  # https://github.com/brentp/somalier/issues/26#issuecomment-543120582
+  if c.proportion_other > 0.1: return -1
   if int(c.nref + c.nalt) < min_depth:
     return -1
   if c.nalt == 0:
     return 0
   result = c.nalt.float / (c.nalt + c.nref).float
 
-proc alts*(ab:float): int8 {.inline.} =
+proc alts*(ab:float, ab_cutoff:float=0.04): int8 {.inline.} =
   if ab < 0: return -1
-  if ab < 0.02: return 0
-  if ab > 0.98: return 2
+  if ab < ab_cutoff: return 0
+  if ab > (1 - ab_cutoff): return 2
   if ab >= 0.2 and ab <= 0.8: return 1
   return -1
 
@@ -464,14 +494,14 @@ specified as comma-separated groups per line e.g.:
 
   if opts.ped != "":
     samples = parse_ped(opts.ped)
-    echo samples.len
-  if samples.len > 30000:
+  if samples.len > 30_000:
     stderr.write_line "[somalier] WARNING!! somalier will work fine for even 100K samples, but it is not optimal for such scenarios."
     stderr.write_line "[somalier] ......... please open an issue at: https://github.com/brentp/somalier/issues as larger cohorts"
     stderr.write_line "[somalier] ......... can be supported."
 
   groups.add_ped_samples(samples, final.samples)
-  groups.add(readGroups(opts.groups))
+  # send in groups so we can adjust baed on self-self samples
+  groups.add(readGroups(opts.groups, groups))
   stderr.write_line &"[somalier] time to get expected relatedness from pedigree graph: {cpuTime() - t0:.2f}"